speed up some processing of student with emphases
now's when we tackle "doing fewer iterations" instead of "do less work per
iteration" lol

okay so it's entirely possible that by rearranging how I treat areas of
emphasis I may have just cut this person's number of iterations to ~421,000

so. An area specification is essentially a tree, right? Each requirement is a
node on the tree.

```yaml
result:
  all:
    - requirement: Core
    - requirement: Musicology
    - requirement: Additional Upper-Level Elective
    - requirement: Additional Electives
    - requirement: Education
    - requirement: Required Solo Performances
    - requirement: Required Recital Attendance
    - requirement: Senior Half Recital
    - requirement: Art, Theater, or Lyric Theater
```

So the way I built emphases was, essentially, the body of the emphasis just
gets appended onto the tree of the main area – becoming, essentially, this:

```diff
 result:
   all:
     - requirement: Core
     - requirement: Musicology
     - requirement: Additional Upper-Level Elective
     - requirement: Additional Electives
     - requirement: Education
     - requirement: Required Solo Performances
     - requirement: Required Recital Attendance
     - requirement: Senior Half Recital
     - requirement: Art, Theater, or Lyric Theater
+    - requirement: Emphasis
```

… where the vocal emphasis itself looks like this:

```yaml
result:
  all:
    - requirement: Core
    - requirement: Electives
    - requirement: Education
    - requirement: Performance Studies
    - requirement: Official Choral Ensemble Participation
    - requirement: Required Vocal Chamber Music Performances
```

So I then have a thing that looks through all of the requirements at the top
level. It finds which requirements are disjoint with respect to all other
reqs, IE have no possible overlapping courses, and then solves them once
instead of putting them in the main solutions method. So if there's a
200-possibility Electives, and a 200-possibility Core, and there's no overlap
between the two, they can solve separately, and there's only one "solution"
generated, instead of 200 x 200 solutions.

So what this meant was, if an emphasis had any overlap with _any_ of the other
top-level requirements, it would not be solvable independently, and would do
something like this:

```
emitting 431,681,536 solutions at $ → .count
	$ → .count → [0] → %Core: 1
	$ → .count → [1] → %Musicology: 2
	$ → .count → [2] → %Additional Upper-Level Elective: 2
	$ → .count → [3] → %Additional Electives: 13
	$ → .count → [9] → %Emphasis: Vocal: 8,301,568
```

1 for Core, 2 for Musicology, 2 for upper-level, 13 for electives… and 8+
million for the Emphasis.

All of the items in that list are then fed into the giant
`itertools.combinations` combinator, which is what drives the whole program.

So, what I just did, is instead of inserting the emphasis as it's own "block"
or "requirement" at the top level, is to assert that an emphasis may only
start with an all-of rule, and then insert each of the emphasis' requirements
as a top-level requirement, instead.

That gives us this model:

```
emitting 421,564 solutions at $ → .count
	$ → .count → [0] → %Core: 1
	$ → .count → [1] → %Musicology: 2
	$ → .count → [2] → %Additional Upper-Level Elective: 2
	$ → .count → [3] → %Additional Electives: 13
	$ → .count → [9] → %Emphasis: Vocal → Core: 1
	$ → .count → [10] → %Emphasis: Vocal → Electives: 8,107
```

which is only 1 x 2 x 2 x 13 x 1 x 8107.

Every solution is still represented there; the disjoint-finder will only solve
requirements that can't affect anything else, and it looks through all of the
solutions for each disjoint requirement to find the first OK / best solution
within that disjoint requirement.