automatically exclude "required" courses from all query rules
A "required" course is one that must always be taken. To show that a
course is required, it - and all of its ancestors – must be within
"all:"-based count rules.

Any time there is a possibility that a course might not be absolutely
required, we don't include it as a required course.

Once we have our set of required clbids, go through each nested rule.
When you find a Query, add the clbids to the new “excluded_clbids”
property on the query.

Now, when a query calls get_data, the first thing it does is remove all
courses with clbids in the excluded set, before it runs the normal
query, so that we automatically remove the courses that can’t possibly
be used there.

-----

Psuedo-code for find_required:

For each rule,
  If it is a count rule,
    If it is an all,
      Call find_required on each child
    Else, skip
  If it is a query, skip
  If it is a course,
    Return the discovered clbid from the transcript, or nothing if there are multiple
  If it is a proficiency,
    If it has a course, return the discovered clbid, or nothing if there are multiple
    Else, skip
  If it is a requirement,
    Call find_required on the child
End