restructure dp.audit
reduces the number of events being emitted

also reduces the amount of memory needed, by removing the list of
duration floats that it was using to compute the average loop duration.

a python float float = 24 bytes in bits => 192 bits
iters = 1,200,000,000

(float * iters) in KB => 28,125,000 KB
(float * iters) in MB => 28,800 MB
(float * iters) in GB => 28.8 GB

that's right, for a 1.2 billion loop audit, we're using 28 gigabytes of
RAM, just to store the audit loop durations.

This should reduce that overhead to, oh, about 24 bytes.