add a "lazy_product" function
for computing products lazily.

itertools.product is eager, and it has to be, because Python can't
assume that iterating an iterable multiple times will be idempotent.

We, however, treat non-idempotent iterations as a bug, so we can use
this lazy_product function and trade a bit of execution time for greatly
reduced memory usage and startup time.

For instance, when running our student who has ~960,000,000
combinations, the startup time required to audit the first solution
drops from ~77s to only ~31s