remove ClaimAttempt struct in favor of just using Claim
this halves the number of objects we create in make_claim, and object
creation is the most expensive part of that.

I originally included ClaimAttempt as a sort of half-baked way to track
what rules had attempted to claim a course, but (a) it only tracked
the claims that happened before it was issued, and (b) the only thing
that's used conflicts_with is the CLI.

Now, I think that exposing conflicts_with is valuable, but it should
come from the authoritative source somehow, RequirementContext's list
of claims, not the claims themselves.

Oh, we also make Claim not be a frozen object, because we just don't
mutate it anywhere. attrs' docs recommend not creating Frozen objects
in a hot loop, if possible, and this is typically the most expensive
function we have.