add WIP optimization of solutions generation
The basic theory here is that, assuming that a spec has no
multicountable rules, we should be able to instantly reject any
potential solution that uses a course more than once in the major.

This commit is a quick-and-dirty implementation of that idea, comparing
a list of courses against the set of unique courses in that list.

It has issues, but we are currently whitelisting it for some students
who take a long time to audit otherwise.