fix issue where negative credit values would be ignored in sum(credits)
this was a botched output optimization to omit 0-credit courses from
the results of `sum(credits)`.

i changed it from `> 0 credits` to `!= 0 credits` to include negative
credit courses.

this is an issue because, in order to allow off-campus study, we
registered everyone with an OFFC course (4-credits) for a
negative-4-credit course to allow them to register anyway.