automatically delete old, identical audits
whenever we insert a new one, look to see if the result string
already exists in the table; if it does, then we should delete
any old audits with that same output, and then insert the new one.

Now that we're not storing hashed values in the json output, we
have many more exact duplicate values, and by comparing the jsonb
text exactly (ignoring sorting, because that's what postgres does),
then we avoid any possibility of hash collisions.