move assertion value exceptions to be loaded at the beginning
instead of once per iteration. Not only is it more efficient, but it
also solves a bug that we ran into - the code that looks for the minimal
solution size would see the non-exception version of the assertion, so
if we had a 1-course assertion, and changed it to 3, the solver would
only return 1-tuple items, instead of 3-tuples.